requirements.txt
requests
pydantic
python-dotenv
httpx
